% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.multi.r
\name{fun.sim}
\alias{fun.sim}
\title{Simulation function of quantitative multigroup data under a
multivariate normal distribution}
\usage{
fun.sim(g, mean1, d, n.var, sds2, corr)
}
\arguments{
\item{g}{An vector of the size of each group}

\item{mean1}{An vector of the population means structure}

\item{d}{distance d for the structure of population means}

\item{n.var}{2x1 dimension vector whose first component is the number of
random variables to simulate and the second component number
of noise variables to simulate}

\item{sds2}{An vector of the variances to simulate for each group noise variables}

\item{corr}{An vector of the correlation to simulate for each group and noise variables}
}
\value{
\strong{return an grafics}
}
\description{
Simulation function of quantitative multigroup data under a
multivariate normal distribution
}
\examples{
fun.sim(g=c(20,20),mean1=2,d=0,sds2=c(1,1,1),corr=c(0.5,0.5,0),n.var=c(50,1))
}
