% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcovmultipca.r
\name{new.cov}
\alias{new.cov}
\title{Function for the new covariance matrix in the multigroup PCA method}
\usage{
new.cov(x, cls, A)
}
\arguments{
\item{x}{is a matrix with the data}

\item{cls}{is a vector of classes}

\item{A}{is a symmetric and positive definite matrix associated to inner product respect to the base of its vectorial space.}
}
\value{
\strong{return an grafics}.
}
\description{
Generates covariance matrix...
}
\examples{
library(plsgenomics)
data(SRBCT)
mydata<-SRBCT$X
mydata<-mydata[1:50,1:20]
groups<-as.factor(SRBCT$Y)[1:50]
new.cov(x=mydata,cls=groups,A=diag(ncol(mydata)))
}
