% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{pca}
\alias{pca}
\title{Performs a principal components analysis in the
given data matrix.
Show PCA graphical output.}
\usage{
pca(datos, grupos, Plot = TRUE, center = TRUE, scale = TRUE)
}
\arguments{
\item{datos}{is a matrix with the data}

\item{grupos}{is a vector of classes}

\item{Plot}{vector logic for grafic}

\item{center}{data set center by columns}

\item{scale}{data set scaled by columns}
}
\value{
\strong{return an grafics}.
}
\description{
Performs a principal components analysis in the
given data matrix.
Show PCA graphical output.
}
\examples{
library(plsgenomics)
data(SRBCT)
mydata<-SRBCT$X
mydata<-mydata[1:30,1:20]
groups<-as.factor(SRBCT$Y)[1:30]
pca(datos=mydata,grupos=groups,Plot=TRUE,center=TRUE,scale=TRUE)
}
